# docs/swagger.yaml
definitions:
  dispute.CloseDisputeRequest:
    properties:
      dispute_id:
        type: integer
    required:
    - dispute_id
    type: object
  dispute.CreateDisputeRequest:
    properties:
      order_id:
        type: integer
      reason:
        maxLength: 100
        type: string
    required:
    - order_id
    - reason
    type: object
  dispute.ListDisputesRequest:
    properties:
      dispute_id:
        type: integer
      page:
        minimum: 1
        type: integer
      page_size:
        maximum: 100
        minimum: 1
        type: integer
      status:
        enum:
        - disputing
        - refund
        - closed
        type: string
    type: object
  dispute.RefundReviewRequest:
    properties:
      dispute_id:
        type: integer
      reason:
        maxLength: 100
        type: string
      status:
        enum:
        - refund
        - closed
        type: string
    required:
    - dispute_id
    - status
    type: object
...
  order.TransactionListRequest:
    properties:
      client_id:
        type: string
      endTime:
        type: string
      page:
        minimum: 1
        type: integer
      page_size:
        maximum: 100
        minimum: 1
        type: integer
      startTime:
        type: string
      status:
        enum:
        - success
        - pending
        - failed
        - expired
        - disputing
        - refund
        - refused
        type: string
      type:
        enum:
        - receive
        - payment
        - transfer
        - community
        type: string
    type: object
...
  util.ResponseAny:
    properties:
      data: {}
      error_msg:
        example: ""
        type: string
    type: object
info:
  contact: {}
  title: LINUX DO PAY
  version: 0.1.0
paths:
...
  /api/v1/order/dispute:
    post:
      consumes:
      - application/json
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dispute.CreateDisputeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - order
  /api/v1/order/dispute/close:
    post:
      consumes:
      - application/json
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dispute.CloseDisputeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - order
  /api/v1/order/disputes:
    post:
      consumes:
      - application/json
      parameters:
      - description: request body
        in: body
        name: request
        schema:
          $ref: '#/definitions/dispute.ListDisputesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - order
  /api/v1/order/disputes/merchant:
    post:
      consumes:
      - application/json
      parameters:
      - description: request body
        in: body
        name: request
        schema:
          $ref: '#/definitions/dispute.ListDisputesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - order
  /api/v1/order/refund-review:
    post:
      consumes:
      - application/json
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dispute.RefundReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - order
  /api/v1/order/transactions:
    post:
      consumes:
      - application/json
      parameters:
      - description: request body
        in: body
        name: request
        schema:
          $ref: '#/definitions/order.TransactionListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - order
...
swagger: "2.0"